% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pure_tumour_methylation.R
\name{annotate_copy_number}
\alias{annotate_copy_number}
\title{Assign copy number calls}
\usage{
annotate_copy_number(dt_sample, seg, rm_sex_chrom = FALSE)
}
\arguments{
\item{dt_sample}{data.table object with each CpG and their coverage, counts methylated and methylation rate}

\item{seg}{ASCAT.m copy number segements object}

\item{rm_sex_chrom}{Logical indicating if you would like to remove sex chrom from downstream analyses}
}
\value{
A dataframe for each sample_id with the copy number calls added
}
\description{
\code{annotate_copy_number} returns the data.table dt_sample annotated with allele-specific copy numbers
}
