% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ASCAT.m.R
\name{ascat.m.plotRawData}
\alias{ascat.m.plotRawData}
\title{ascat.m.plotRawData}
\usage{
ascat.m.plotRawData(ASCATobj, raw_LogR, pch, cex, lim_logR)
}
\arguments{
\item{ASCATobj}{an ASCAT object (e.g. data structure from ascat.loadData)}

\item{raw_LogR}{vector with the LogR values before correction}

\item{pch}{type of data points in plot}

\item{cex}{size of data points in plot}

\item{lim_logR}{y-axis limits on logR plot}
}
\value{
Produces png files showing the logR and BAF values for tumour and germline samples
}
\description{
Plot tumour and germline BAF and LogR
}
\author{
Peter Van Loo
}
