% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ASCAT.m.R
\name{ascat.m.plotSegmentedData}
\alias{ascat.m.plotSegmentedData}
\title{ascat.m.plotSegmentedData}
\usage{
ascat.m.plotSegmentedData(ASCATobj, lim_logR = 2)
}
\arguments{
\item{ASCATobj}{an ASCAT object (e.g. data structure from ascat.loadData)}
}
\value{
Produces png files showing the logR and BAF values for tumour and germline samples
}
\description{
Plot segmentated BAF LogR
}
\author{
Peter Van Loo
}
