% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ascat.plotRawData.flags}
\alias{ascat.plotRawData.flags}
\title{ascat.plotRawData}
\usage{
ascat.plotRawData.flags(ASCATobj, pch, cex, lim_logR)
}
\arguments{
\item{ASCATobj}{an ASCAT object (e.g. data structure from ascat.loadData)}

\item{pch}{type of data points in plot}

\item{cex}{size of data points in plot}

\item{lim_logR}{y-axis limits on logR plot}
}
\value{
Produces png files showing the logR and BAF values for tumour and germline samples
}
\description{
Plot BAF LogR
}
\author{
Peter Van Loo
}
\keyword{internal}
