% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_differential_methylation.R
\name{bin_CpGs}
\alias{bin_CpGs}
\title{Cluster CpGs into annotated bins}
\usage{
bin_CpGs(path, patient_id, sample_id, dt, anno_list, n_cores)
}
\arguments{
\item{path}{Character string of the output directory}

\item{patient_id}{Character string containting the patient ID}

\item{sample_id}{Character string containting the sample ID.}

\item{dt}{data.table where each CG is a row with DMP info.}

\item{anno_list}{A data.table object containing annotated genomic bins including
genes, exons, introns, UTRs, CGI, CGI shores, CGI shelves, promoters or enhancers}

\item{n_cores}{number of cores for parallel processing}
}
\value{
A dataframe for each sample_id with the copy number calls added
}
\description{
\code{bin_CpGs} returns the df with the annotation for each CpG
}
