% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_wgbs_methylation.R
\name{calculate_m_t_hdi}
\alias{calculate_m_t_hdi}
\title{Calculate HDI by simulation}
\usage{
calculate_m_t_hdi(meth_c, n_cores, itersplit = 100000)
}
\description{
Calculate HDI by simulation
}
\keyword{internal}
