% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_wgbs_differential_meth.R
\name{call_dmps}
\alias{call_dmps}
\title{Call DMPs returns the CAMDAC DMP dataframe
Call differentially methylated positions}
\usage{
call_dmps(
  pmeth,
  nmeth,
  effect_size = 0.2,
  prob = 0.99,
  itersplit = 500000,
  ncores = 5
)
}
\description{
Call DMPs returns the CAMDAC DMP dataframe
Call differentially methylated positions
}
\keyword{internal}
