% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_wgbs_differential_meth.R
\name{call_dmr_routine}
\alias{call_dmr_routine}
\title{Function to call DMRs on a camdac dmp dataset}
\usage{
call_dmr_routine(
  tmeth_dmps,
  regions_annotations,
  min_DMP_counts,
  min_consec_DMP
)
}
\description{
Function to call DMRs on a camdac dmp dataset
}
\keyword{internal}
