% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_wgbs_battenberg.R
\name{camdac_to_battenberg_prepare_wgbs}
\alias{camdac_to_battenberg_prepare_wgbs}
\title{Generate alleleCounter file from CAMDAC}
\usage{
camdac_to_battenberg_prepare_wgbs(
  tumour_prefix,
  normal_prefix,
  camdac_tsnps,
  outdir
)
}
\arguments{
\item{camdac_tumour_ac}{CAMDAC tumour allele counts filepath. Expected *.gz}

\item{camdac_normal_ac}{CAMDAC normal allele couts filepath. Expected *.gz}

\item{camdac_tnsps}{CAMDAC tumour-normal-snps object. Expected *.gz}

\item{output_file}{allelecounter formatted-file output directory.}
}
\value{
File handle for allele counter file generated
}
\description{
\code{camdac_to_battenberg_prepare_wgbs} converts CAMDAC allele counter results to a format for processing.
}
\keyword{internal}
