% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmain_pipeline.R
\name{cmain_call_dmps}
\alias{cmain_call_dmps}
\title{Call tumor-normal DMPs}
\usage{
cmain_call_dmps(tumour, normal, config)
}
\arguments{
\item{normal}{A camdac sample object}

\item{config}{A camdac config object}

\item{tumor}{A camdac sample object}
}
\description{
Single-sample DMP calling on CAMDAC-deconvolved data
}
