% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmain_pipeline.R
\name{cmain_call_dmrs}
\alias{cmain_call_dmrs}
\title{Call tumor-normal DMRs}
\usage{
cmain_call_dmrs(tumour, config)
}
\arguments{
\item{config}{A camdac config object}

\item{tumor}{A camdac sample object}

\item{normal}{A camdac sample object}
}
\description{
Single-sample DMR calling on CAMDAC DMP data
}
