% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmain_pipeline.R
\name{cmain_count_alleles}
\alias{cmain_count_alleles}
\title{Count alleles}
\usage{
cmain_count_alleles(sample, config)
}
\arguments{
\item{sample}{A camdac sample object}

\item{config}{A camac allele object}
}
\description{
Count alleles
}
