% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmain_pipeline.R
\name{cmain_deconvolve_methylation}
\alias{cmain_deconvolve_methylation}
\title{Deconvolve methylation}
\usage{
cmain_deconvolve_methylation(tumour, normal, config)
}
\arguments{
\item{normal}{A camdac sample object}

\item{config}{A camdac config object}

\item{tumor}{A camdac sample object}
}
\description{
Deconvolve methylation
}
