% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmain_pipeline.R
\name{cmain_make_methylation_profile}
\alias{cmain_make_methylation_profile}
\title{Make methylation}
\usage{
cmain_make_methylation_profile(sample, config)
}
\arguments{
\item{sample}{A camdac sample object}

\item{config}{A camdac config object}
}
\description{
Pre-process methylation from allele counts for CAMDAC deconvolution
}
