% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmain_pipeline.R
\name{cmain_make_snp_profiles}
\alias{cmain_make_snp_profiles}
\title{Make SNPs}
\usage{
cmain_make_snp_profiles(tumour, normal, config)
}
\arguments{
\item{normal}{A camdac sample object}

\item{config}{A camdac config object}

\item{tumor}{A camdac sample object}
}
\description{
Format and save SNP file for CNA analysis (ASCAT or BATTENBERG)
}
