% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmain_pipeline.R
\name{cmain_run_battenberg}
\alias{cmain_run_battenberg}
\title{Run battenberg}
\usage{
cmain_run_battenberg(tumour, normal, config)
}
\arguments{
\item{normal}{A camdac sample object}

\item{config}{A camdac config object}

\item{tumor}{A camdac sample object}
}
\description{
Expects SNP profiles to have been created using \code{cmain_make_snp_profiles}
}
