% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pure_tumour_methylation.R
\name{compute_tumour_methylome}
\alias{compute_tumour_methylome}
\title{Compute the tumour methylation rate}
\usage{
compute_tumour_methylome(dt, p, min_cov_t = 3, sex, build)
}
\arguments{
\item{dt}{data.table object with each CpG and their coverage, counts methylated,
methylation rate and copy number and matched normal methylation info}

\item{p}{Numerical - Sample purity estimates}

\item{min_cov_t}{Numerical - Minimum tumour coverage}

\item{sex}{Character variable with  the patient expressed as "XX" for female or "XY" for male.}

\item{build}{Character variable corresponding to the reference genome used for alignment.}
}
\value{
A dataframe for each sample_id with the tumour methylome added
}
\description{
\code{compute_tumour_methylome} returns the data.table dt annotated with
CAMDAC pure tumour methylation rates
}
