% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_config.R
\name{create_camdac_sample}
\alias{create_camdac_sample}
\title{Build CAMDAC sample object}
\usage{
create_camdac_sample(patient_id, patient_sex, sample_id, sample_type, bam_file)
}
\arguments{
\item{patient_id}{An identifier for the patient}

\item{patient_sex}{The sex of the patient, "XX" or "XY". Required for CNA calling.}

\item{sample_type}{Currently unused field.}

\item{bam_file}{Sample BAM file}
}
\description{
Build CAMDAC sample object
}
