% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_config.R
\name{download_pipeline_files}
\alias{download_pipeline_files}
\title{Download CAMDAC pipeline files}
\usage{
download_pipeline_files(bsseq, directory = NULL)
}
\arguments{
\item{directory}{Optional. Directory to download files to.}

\item{assay}{Sequencing assay. Either wgbs or rrbs.}
}
\description{
CAMDAC pipeline files are required for analysis. This function downloads the files to
the output directory and unpacks them. By default, CAMDAC searches for the files in the
environment variable CAMDAC_PIPELINE_FILES. If this is missing, the current directory is used.
}
