% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_methylation_data_processing.R
\name{format_methylation_df}
\alias{format_methylation_df}
\title{Format methylation rates
\code{format_methylation_df}}
\usage{
format_methylation_df(
  dt,
  sample_id,
  normal_ids,
  path_output,
  n_cores,
  suffix,
  trim = FALSE
)
}
\arguments{
\item{dt}{data.table containing the methylation information for each CpG}

\item{sample_id}{sample ID}

\item{normal_ids}{sample ID of normal sample(s)}

\item{path_output}{output directory}

\item{n_cores}{number of threads for HDI calculation}

\item{suffix}{string containing the column names suffix for normal samples
This is to distinguish between the proxy supplied for the normal infiltrates
for use in deconvolution and the normal cell of origin for use in DMP/DMR calling}

\item{trim}{Logical value establishing whether regions with extremely high coverage be trimmed or not}
}
\value{
A GRanges object with all the CpG loci, their coverage, counts methylated and methylation rate
}
\description{
Format methylation rates
\code{format_methylation_df}
}
