% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_output}
\alias{format_output}
\title{Format output nucleotide counts
\code{format_output}}
\usage{
format_output(
  patient_id,
  sample_id,
  sex,
  is_normal = FALSE,
  path,
  path_to_CAMDAC,
  build,
  txt_output = FALSE
)
}
\arguments{
\item{patient_id}{Character variable containting the patient id number}

\item{sample_id}{Character variable with the sample ID}

\item{sex}{Character variable with  the patient expressed as "XX" for female or "XY" for male.}

\item{is_normal}{Logical flag set to false if the sample to be formatted is normal or tumour}

\item{path}{Character path variable pointing to the desired working directory.
This is where the output will be stored and should be constant for all CAMDAC functions.
Do not alter the output directory structure while running CAMDAC.}

\item{path_to_CAMDAC}{Character variable containting the path to the CAMDAC directory
including dir name  (e.g. "/path/to/CAMDAC/").}

\item{build}{Character variable corresponding to the reference genome used for alignment.
CAMDAC is compatible with "hg19", "hg38", "GRCH37","GRCH38".}

\item{txt_output}{Character varaiable indicating if .txt output
is desired in addition to GRanges object in .RData file}
}
\value{
Concatenated SNP and CpG information
}
\description{
Format output nucleotide counts
\code{format_output}
}
