% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_wgbs_differential_meth.R
\name{get_cluster_counts}
\alias{get_cluster_counts}
\title{Count CpGs within DMP annotations}
\usage{
get_cluster_counts(dt)
}
\description{
Count CpGs within DMP annotations
}
\keyword{internal}
