% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_differential_methylation.R
\name{get_differential_methylation}
\alias{get_differential_methylation}
\title{Perform differential methylation analysis on deconvolute tumour methylation rates}
\usage{
get_differential_methylation(
  patient_id,
  sample_id,
  sex,
  normal_origin_proxy_id,
  path,
  path_to_CAMDAC,
  build,
  effect_size = 0.2,
  prob = 0.99,
  min_DMP_counts_in_DMR = 5,
  min_consec_DMP_in_DMR = 4,
  n_cores,
  reseg = FALSE,
  bulk = FALSE
)
}
\arguments{
\item{patient_id}{Character variable containting the patient id number}

\item{sample_id}{Character variable with the tumour sample_id}

\item{sex}{Character variable with  the patient expressed as "XX" for
female or "XY" for male.}

\item{normal_origin_proxy_id}{Character variable with the sample ID
of the normal to be used as a proxy for the tumour cell of origin in}

\item{path}{Character path variable pointing to the desired working
directory. This is where the output will be stored.}

\item{path_to_CAMDAC}{Character variable containting the path to the CAMDAC
directory including dir name (e.g. "/path/to/CAMDAC/").}

\item{build}{Character variable corresponding to the reference genome
used for alignment. CAMDAC is compatible with "hg19", "hg38", "GRCH37","GRCH38".}

\item{effect_size}{Numerical containting the minimum tumour-normal methylation
difference (default is 0.2)}

\item{prob}{Numerical  value representing the threshold for statistically
significant DMP (default is p=0.99)}

\item{min_DMP_counts_in_DMR}{Numerical value representing the number of
DMPs required in a DMR}

\item{min_consec_DMP_in_DMR}{Numerical value representing the number of
consecutive DMPs required in a DMR}

\item{n_cores}{Numerical value correspdonding to the number of cores
for parallel processing}

\item{reseg}{Logical value should be set to FALSE. Multi-sample re-segmentation of
the copy number profiles will be available in future versions of CAMDAC.}

\item{bulk}{Default is FALSE unless you want bulk DMP/DMR calls in addition
to CAMDAC pure tumour differential methylation analysis

Note:
#' Annotation include:
CGI (including shore and shelves)
gene body (intragenic, 5UTR, 3UTR, intron, exon)
promoter (2kb upstream and 500 downstream any UCSC annotated gene)
enhancer (vista and FANTOM5 annotation)}
}
\value{
Biologically significant DMPs, DMRs
}
\description{
\code{get_differential_methylation}
}
