% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_differential_methylation.R
\name{get_DMPs}
\alias{get_DMPs}
\title{Get DMPs}
\usage{
get_DMPs(path, patient_id, sample_id, df, prob = 0.99, n_cores)
}
\arguments{
\item{path}{Complete path to the CAMDAC methylation output directory
fir this sample}

\item{patient_id}{Character string containting the patient number}

\item{sample_id}{Character variable with the tumour sample_id}

\item{df}{A data.table with pure, bulk and normal methylation info}

\item{prob}{Numerical  value representing the threshold for statistically
significant DMP (default is p=0.99)}

\item{n_cores}{Number of cores to do the statistical testing over}
}
\value{
A data.table object with all the CpG loci, their coverage, counts
methylated and methylation rate
}
\description{
\code{get_DMPs} returns a df with annotated statistics for each CpG
}
