% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_differential_methylation.R
\name{get_DMRs}
\alias{get_DMRs}
\title{Assign bins}
\usage{
get_DMRs(
  path,
  patient_id,
  sample_id,
  dt,
  anno_list,
  min_DMP_counts,
  min_consec_DMP,
  n_cores,
  bulk = FALSE
)
}
\arguments{
\item{path}{Character string of the output directory}

\item{patient_id}{Character string containting the patient_id ID}

\item{sample_id}{Character string containting the sample ID.}

\item{dt}{dataframe where each CG is a row with DMP info.}

\item{anno_list}{A data.table object containing annotated genomic bins including
genes, exons, introns, UTRs, CGI, CGI shores, CGI shelves, promoters or enhancers}

\item{min_DMP_counts}{Numerical - number of DMPs required in a DMR}

\item{min_consec_DMP}{Numerical - number of consecutive DMPs required in a DMR}

\item{n_cores}{number of cores for parallel processing}
}
\value{
A dataframe for each sample_id with the copy number calls added
}
\description{
\code{annotate_DMRs} returns the df with the annotation for each CpG
}
