% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{get_msp1_fragments}
\alias{get_msp1_fragments}
\title{get_msp1_fragments}
\usage{
get_msp1_fragments(dt, build, path_to_CAMDAC, outfile)
}
\arguments{
\item{dt}{data.table object with containing all covered CCGGs in the sample}

\item{build}{Character, Either "hg19", "hg38", "GRCH37","GRCH38"}

\item{path_to_CAMDAC}{Character string containting the path to the CAMDAC dir including
dir name e.g. "~/CAMDAC/"}

\item{outfile}{character srting with output filename}
}
\description{
get msp1 fragments
}
\author{
elizabeth larose cadieux
}
