% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pure_tumour_methylation.R
\name{get_pure_tumour_methylation}
\alias{get_pure_tumour_methylation}
\title{Deconvolve the pure tumour methylation rate from bisulfite sequencing data}
\usage{
get_pure_tumour_methylation(
  patient_id,
  sample_id,
  sex,
  normal_infiltrates_proxy_id,
  path,
  path_to_CAMDAC,
  build,
  n_cores,
  reseg = FALSE
)
}
\arguments{
\item{patient_id}{Character variable containting the patient id number}

\item{sample_id}{Character variable with the (control or tumour) sample_id}

\item{sex}{Character variable with  the patient expressed as "XX" for
female or "XY" for male.}

\item{normal_infiltrates_proxy_id, }{Sample ID of the matched normal control}

\item{path}{Character path variable pointing to the desired working directory.
This is where the output will be stored and should be constant for all CAMDAC functions.}

\item{path_to_CAMDAC}{Character variable containting the path to the CAMDAC
directory including dir name (e.g. "/path/to/CAMDAC/").}

\item{build}{Character variable corresponding to the reference genome
used for alignment. CAMDAC is compatible with "hg19", "hg38", "GRCH37","GRCH38".}

\item{n_cores}{Numerical value correspdonding to the number of cores
for parallel processing}

\item{reseg}{Logical value should be set to FALSE. Multi-sample re-segmentation of
the copy number profiles will be available in future versions of CAMDAC.

Note:
#' Annotation include:
CGI (including shore and shelves)
gene body (intragenic, 5UTR, 3UTR, intron, exon)
promoter (2kb upstream and 500 downstream any UCSC annotated gene)
enhancer (vista and FANTOM5 annotation)}
}
\value{
CAMDAC purified tumour methylation rates
}
\description{
\code{get_pure_tumour_methylation}
}
