% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_config.R
\name{get_reference_files}
\alias{get_reference_files}
\title{Get CAMDAC reference files from config}
\usage{
get_reference_files(config, type_folder, glob = NULL)
}
\description{
Get CAMDAC reference files from config
}
