% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_wgbs_methylation.R
\name{HDIofICDF}
\alias{HDIofICDF}
\title{HDI of ICDF}
\usage{
HDIofICDF(ICDFname, credMass = 0.99, tol = 0.0001, ...)
}
\arguments{
\item{ICDFname}{The inverse cumulative density function of the distribution.}

\item{credMass}{The desired mass of the HDI region.}

\item{tol}{Tolerance parameter for optimisation. the lower the tolerance,the
longer the optimisation, but the higher the accuracy.
According to CAMDAC RRBS comments, tol=1e-4 gives values
of the same accuracy as our max resolution.
This function is adapted from Greg Snow's TeachingDemos package
E.g.Determine HDI of a M=30 and UM=12 CpG
Adding 1 to shape parameter ensures uniform beta(1,1) is updated with our counts
HDIofICDF(qbeta,shape1 = 30+1 , shape2 = 12+1 )}
}
\value{
Highest density interval (HDI) limits in a vector.
}
\description{
HDI of ICDF
}
\keyword{internal}
