% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_wgbs_methylation.R
\name{HDIofMCMC_mt}
\alias{HDIofMCMC_mt}
\title{Calculate HDI by simulation}
\usage{
HDIofMCMC_mt(M_b, UM_b, M_n, UM_n, p, CN, credMass = 0.99)
}
\arguments{
\item{M_b}{counts methylated in the tumour}

\item{UM_b}{counts unmethylated in the tumour}

\item{M_n}{counts methylated in the normal}

\item{UM_n}{counts unmethylated in the normal}

\item{p}{tumour purity}

\item{CN}{total tumour copy number}

\item{credMass}{default is 0.99
credMass is a scalar between 0 and 1, indicating the mass within the
credible interval that is to be estimated.}

\item{CN_n}{total normal copy number}
}
\value{
Value: HDIlim is a vector containing the limits of the HDI
}
\description{
Computes highest density interval from a sample of representative values,
estimated as shortest credible interval for a unimodal distribution
}
\keyword{internal}
