% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_helpers.R
\name{helper_camdac_pileup}
\alias{helper_camdac_pileup}
\title{Cache existing CAMDAC results into a sub-directory so that the current ones can be
overwritten by the refitting pipeline
Decided this is unnecessary as the initial results were so wrong.
Exported only for development}
\usage{
helper_camdac_pileup(bam_file, seg, loci_dt)
}
\description{
Cache existing CAMDAC results into a sub-directory so that the current ones can be
overwritten by the refitting pipeline
Decided this is unnecessary as the initial results were so wrong.
Exported only for development
}
\keyword{internal}
