% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_helpers.R
\name{load_cna_data}
\alias{load_cna_data}
\title{Parse ASCAT and bb output directories to load CNA data}
\usage{
load_cna_data(tumour, config, data_type)
}
\description{
See "annotate_copy_number" func
A function required to load copy number for a tumour sample from camdac, either ascat or bb,
result should be: chrom, start, end, nA, nB, CN (total), seg_min and seg_max.
This should also include the purity and ploidy. As a separate list?
note that seg_min and seg_max are actually duplicates of the start and end columns, required to
keep track of the ascat segment positions after overalp
WARN: This drops sex chromosome but not implimented. Also should drops CN=0 (hom del) regions
}
