% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ASCAT.m.R
\name{LogR_correction}
\alias{LogR_correction}
\title{LogR_correction}
\usage{
LogR_correction(
  dt_sample,
  dt_SNPs,
  build,
  chr_names,
  min_normal,
  fragments_file,
  replic_timing_file_prefix,
  n_cores
)
}
\arguments{
\item{dt_sample}{Allelecounts output as a data.table}

\item{dt_SNPs}{Allelecounts output subset to QC'ed SNP positions}

\item{build}{Character variable corresponding to the reference genome version used for alignment}

\item{chr_names}{Character variable with the seqlevels.}

\item{min_normal}{Numerical with the minimum normal coverage threshold}

\item{fragments_file}{CAMDAC reference MspI fragments file}

\item{replic_timing_file_prefix}{CAMDAC reference replication timing files path and file name prefix}

\item{n_cores}{Numerical value correspdonding to the number of cores for parallel processing}
}
\description{
Correct logR for msp1 fragment size bias and GC content
}
\author{
Elizabeth Larose Cadieux
}
