% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipeline_tumor_normal}
\alias{pipeline_tumor_normal}
\title{CAMDAC tumor-normal pipeline}
\usage{
pipeline_tumor_normal(tumor, normal, config)
}
\arguments{
\item{tumor.}{Tumor sample data built with \code{create_camdac_sample()}.}

\item{normal.}{Normal sample data built with \code{create_camdac_sample()}.}

\item{config.}{Configuration built with \code{create_camdac_config()}.}
}
\description{
Run CAMDAC analysis on a bulk tumor and patient-matched tissue-matched tumor-adjacent normal sample.
}
