% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ASCAT.m.R
\name{plot_BAF_and_LogR}
\alias{plot_BAF_and_LogR}
\title{Plot BAF and logR profiles with ggplot}
\usage{
plot_BAF_and_LogR(dt, outfile, downsample = 100000)
}
\arguments{
\item{dt}{data.frame with methylation info}

\item{outfile}{character srting with output pdf filename
Saves a pdf w/ methylation rate distribution, biases at polymorphic and
non-polymorphic CG/CCGG and coverage distribution}
}
\description{
Plot BAF and logR profiles with ggplot
}
\author{
Elizabeth Larose Cadieux
}
