% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_methylation_data_processing.R
\name{plot_methylation_info}
\alias{plot_methylation_info}
\title{Plot Methylation}
\usage{
plot_methylation_info(df_sample, outfile)
}
\arguments{
\item{df_sample}{data.frame with methylation info}

\item{outfile}{character srting with output pdf filename}

\item{dt}{Data.table that the grob will be made out of}

\item{title_v}{Title for display}

\item{fontsize_v}{Fontsize for title. Default is 14 (goes well with my_theme)}
}
\value{
pdf w/ methylation rate distribution, biases at polymorphic and non-polymorphic CG/CCGG and coverage distribution
Create custom table grob with title
}
\description{
Creates table grob in format that is most common for my usage.
}
\details{
\code{plot_methylation_info} returns the df_sample with annotated q-value for each CpG
}
