% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ASCAT.m.R
\name{plot_normal_SNP_info}
\alias{plot_normal_SNP_info}
\title{Plot plots SNP QC}
\usage{
plot_normal_SNP_info(dt, outfile, min)
}
\arguments{
\item{dt}{data.table with SNP info}

\item{outfile}{character srting with output pdf filename}
}
\value{
pdf
}
\description{
Plot plots SNP QC
}
\author{
Elizabeth Larose Cadieux
}
