% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ASCAT.m.R
\name{plot_SNP_info}
\alias{plot_SNP_info}
\title{Plot SNP data summary and QC}
\usage{
plot_SNP_info(dt, outfile, min)
}
\arguments{
\item{dt}{data.table with SNP info}

\item{outfile}{character srting with output pdf filename}
}
\value{
pdf
}
\description{
\code{plot_SNP_info} plots SNP QC
}
\author{
Elizabeth Larose Cadieux
}
