% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ASCAT.m.R
\name{remove_low_cov_singletons}
\alias{remove_low_cov_singletons}
\title{remove_low_cov_singletons}
\usage{
remove_low_cov_singletons(dt_sample_SNPs, min)
}
\description{
Remove low coverage singletons outliers
}
\author{
Elizabeth larose cadieux
}
