% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ASCAT.m.R
\name{run_ASCAT.m}
\alias{run_ASCAT.m}
\title{Obtain allele-specific copy number profiles, tumour purity and plot SNP data}
\usage{
run_ASCAT.m(
  patient_id,
  sample_id,
  sex,
  patient_matched_normal_id = NULL,
  path,
  path_to_CAMDAC,
  build,
  min_normal = 10,
  min_tumour = 1,
  n_cores = 1,
  reference_panel_coverage = NULL
)
}
\arguments{
\item{patient_id}{Character variable containting the patient id number}

\item{sample_id}{Character variable with the (control or tumour) sample_id}

\item{sex}{Character variable with  the patient expressed as "XX" for female
or "XY" for male.
This is important for copy number profiling. If sex is unknown, put "XY" for now,
then look at the allelic imbalance (BAF) on X in the germline outside pseudo-
autosomal regions. If there are little to no heterozygous SNPs, the sample is likely male.}

\item{patient_matched_normal_id}{Character variable with the sample ID of the matched normal control}

\item{path}{Character path variable pointing to the desired working directory.
This is where the output will be stored
IMPORTANT: The function output directory will be the in the path variable working
directory under "./Copy_number/sample_id/".}

\item{path_to_CAMDAC}{Character variable containting the path to the CAMDAC dir
including dir name (e.g. "/path/to/CAMDAC/").}

\item{build}{Character variable corresponding to the reference genome used for alignment.
CAMDAC is compatible with "hg19", "hg38", "GRCH37","GRCH38".}

\item{min_normal}{Numerical value correspdonding to the minimum counts for germline
SNPs to be included (default:1)}

\item{min_tumour}{Numerical value correspdonding to the minimum counts in the tumour
sample for germline SNPs to be included (default:10)}

\item{n_cores}{Numerical value correspdonding to the number of cores for parallel processing}

\item{reference_panel_coverage}{Path to the reference panel for the coverage.}
}
\value{
Three text files with all the CpG loci and their SNP and/or CpG methylation info
}
\description{
\code{run_ASCAT.m}
}
