% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdac_config.R
\name{split_segments_gr}
\alias{split_segments_gr}
\title{Split genome into segments for allele counting}
\usage{
split_segments_gr(segments_file, n_seg_split)
}
\arguments{
\item{segments_file}{An RDS file containing a GRanges object with each chromosome region to split}

\item{n_seg_split}{An integer to split each chromosome segment}
}
\description{
Split genome into segments for allele counting
}
\keyword{internal}
