% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_tumor_normal.R
\name{pipeline_tumor_normal}
\alias{pipeline_tumor_normal}
\title{Call CAMDAC for a tumor and patient-matched normal sample}
\usage{
pipeline_tumor_normal(
  patient_id,
  tumor_id,
  normal_id,
  tumor_bam,
  normal_bam,
  sex,
  path,
  pipeline_files,
  build,
  min_tumor = 3,
  min_normal = 10,
  n_cores = 1,
  mq = 0
)
}
\arguments{
\item{patient_id}{character. Patient identifier}

\item{tumor_id}{character. Tumor sample identifier}

\item{normal_id}{character. Normal sample identifier}

\item{tumor_bam}{character. Full path to tumor bam file}

\item{normal_bam}{character. Full path to normal bam file}

\item{sex}{character. Patient sex: "XX" for female or "XY" for male}

\item{path}{character. Full path to CAMDAC output directory}

\item{pipeline_files}{character. Full path to parent directory containing CAMDAC pipeline_files}

\item{build}{character. Genome build: "hg19" or "hg38"}

\item{min_tumor}{integer. Minimum read filter for tumor samples}

\item{min_normal}{integer. Minimum read filter for normal samples}

\item{n_cores}{integer. Number of cores to use for parallel processing}

\item{mq}{integer. Minimum mapping quality filter}
}
\description{
Call CAMDAC for a tumor and patient-matched normal sample
}
